% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agreem_binary}
\alias{agreem_binary}
\title{Three reliability studies for some binary rating}
\format{
A list that contains three matrices. Each matrix contains the result of a
study involving two raters. It is a binary rating scale ("+" and "-").
}
\source{
Chapter 18, Problems 18.3
}
\usage{
agreem_binary
}
\description{
The data are reported in a textbook from Fleiss, probably it is fictitious.
}
\references{
Fleiss, J. L., Levin, B., & Paik, M. C. Statistical
Methods for Rates and Proportions, 3rd edition, 2003, ISBN 0-471-52629-0
}
\keyword{datasets}
