% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depression}
\alias{depression}
\title{Depression screening}
\format{
A matrix with 50 observations and 3 variables:
\describe{
\item{depression}{diagnoses as determined by the Cutoff program}
\item{BDI}{Beck Depression Inventory}
\item{GHQ}{General Health Questionnaire}
}
}
\usage{
depression
}
\description{
Fifty general hospital patients, admitted to the Monash Medical Centre in
Melbourne, were randomly drawn from a larger sample described by Clarke et
al. (1993). Agreement between two different screening tests and a diagnosis
of depression was compared. Definition of depression included DSM-III-R Major
Depression, Dysthymia, Adjustment Disorderwith Depressed Mood, and Depression
NOS. Depression was determined empirically using the Cutoff (McKenzie &
Clarke, 1992) program.
The screening tests consisted of
}
\details{
\enumerate{
\item the Beck Depression Inventory (BDI) (Beck et al., 1961) and
\item the General Health Questionnaire (GHQ) (Goldberg & Williams, 1988)
}
}
\references{
McKenzie, D. P. et al., Comparing Correlated Kappas by Resampling:
Is One Level of Agreement Significantly Different from Another?
J. psychiat. Res, Vol. 30, 1996. \doi{10.1016/S0022-3956(96)00033-7}
}
\keyword{datasets}
