% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaGold-package.R
\docType{package}
\name{kappaGold}
\alias{kappaGold-package}
\alias{kappaGold}
\title{kappaGold package}
\description{
Estimate agreement with a gold-standard rating for nominal categories.
}
\author{
\strong{Maintainer}: Matthias Kuhn \email{matthias.kuhn@tu-dresden.de} (\href{https://orcid.org/0000-0003-2868-5155}{ORCID})

Authors:
\itemize{
  \item Jonas Breidenstein \email{jonas.breidenstein@tu-dresden.de}
}

}
