% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{victorinox}
\alias{victorinox}
\title{delete-1 jackknife estimator}
\usage{
victorinox(est, idx)
}
\arguments{
\item{est}{estimator function}

\item{idx}{maximal index vector for data of estimator}
}
\value{
list with jackknife information, bias and SE
}
\description{
Quick simple jackknife routine to estimate bias and standard error of an
estimator.
}
\references{
https://de.wikipedia.org/wiki/Jackknife-Methode
}
